+++ src/test/scala/com/bisnode/kafka/authorization/OpaAuthorizerSpec.scala
@@ -4,23 +4,41 @@ import java.net.{InetAddress, URI}
 import java.net.http.HttpRequest.BodyPublishers
 import java.net.http.HttpResponse.BodyHandlers
 import java.net.http.{HttpClient, HttpRequest, HttpResponse}
+import java.util.Collections
+import java.{util => ju}
 
+import org.junit.runner.RunWith
+import org.scalatestplus.junit.JUnitRunner
+import com.typesafe.scalalogging.LazyLogging
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.scala.{DefaultScalaModule, ScalaObjectMapper}
+import kafka.network.RequestChannel
 import kafka.network.RequestChannel.Session
-import kafka.security.auth.{Operation, Read, Resource, Topic, Write}
+import org.apache.kafka.common.acl.AclOperation
 import org.apache.kafka.common.resource.PatternType
-import org.apache.kafka.common.security.auth.KafkaPrincipal
-import org.junit.runner.RunWith
+import org.apache.kafka.common.security.auth.{KafkaPrincipal, SecurityProtocol}
+import org.apache.kafka.common.resource.ResourcePattern
+import org.apache.kafka.common.resource.PatternType
+import org.apache.kafka.common.resource.ResourceType
+import org.apache.kafka.server.authorizer.{Action, AuthorizationResult}
 import org.scalatest._
-import org.scalatest.junit.JUnitRunner
+import matchers.should._
+import flatspec._
+import org.apache.kafka.common.ClusterResource
+import org.apache.kafka.common.Endpoint
+import scala.compat.java8.FutureConverters
+import io.netty.util.concurrent.Future
+import scala.util.{Success, Failure}
+import scala.concurrent.ExecutionContext.Implicits.global
+
+import scala.jdk.CollectionConverters._
 
 /**
  * Integration level tests of Kafka policy as described in src/main/rego/README.md
  * NOTE: Requires a running OPA instance with the provided policy loaded
  */
 @RunWith(classOf[JUnitRunner])
-class OpaAuthorizerSpec extends FlatSpec with Matchers with PrivateMethodTester {
+class OpaAuthorizerSpec extends AnyFlatSpec with Matchers with PrivateMethodTester with LazyLogging {
 
   private val opaUrl = "http://localhost:8181/v1/data/kafka/authz/allow"
   private val objectMapper = (new ObjectMapper() with ScalaObjectMapper).registerModule(DefaultScalaModule)
@@ -37,63 +55,104 @@ class OpaAuthorizerSpec extends FlatSpec with Matchers with PrivateMethodTester 
   }
 
   "Request object" should "serialize to JSON" in {
-    val request = createRequest("bob", "bob-topic", Read)
+    val actions = List(
+      createAction("bob-topic", AclOperation.READ),
+    )
+    val request = createRequest("bob", actions)
 
     objectMapper.writeValueAsString(request) should not be "{}"
+    objectMapper.writeValueAsString(request.actions.asJava) should not be "{}"
+    objectMapper.writeValueAsString(request.requestContext) should not be "{}"
   }
 
   "OpaAuthorizer" should "authorize when username matches name of topic" in {
     val opaAuthorizer = setupAuthorizer()
-    val request = createRequest("alice-producer", "alice-topic", Write)
-    val input = request.input
+    val actions = List(
+      createAction("alice-topic", AclOperation.WRITE),
+    )
+    val request = createRequest("alice-producer", actions)
 
-    opaAuthorizer.authorize(input.session, input.operation, input.resource) should be (true)
+    opaAuthorizer.authorize(request.requestContext, request.actions.asJava) should be (List(AuthorizationResult.ALLOWED).asJava)
     opaAuthorizer.getCache.size should be (1)
+  }
+
+  "OpaAuthorizer" should "return authorization results for multiple actions in the same request in right order" in {
+    val opaAuthorizer = setupAuthorizer()
+    val actions = List(
+      createAction("alice-topic", AclOperation.WRITE),
+      createAction("alice-topic", AclOperation.WRITE),
+      createAction("alice-topic", AclOperation.READ),
+      createAction("alice-topic", AclOperation.READ),
+      createAction("alice-topic", AclOperation.WRITE),
+      createAction("alice-topic", AclOperation.DESCRIBE),
+      createAction("alice-topic", AclOperation.CREATE),
+    )
+    val request = createRequest("alice-producer", actions)
+
+    opaAuthorizer.authorize(request.requestContext, request.actions.asJava) should be (List(
+      AuthorizationResult.ALLOWED,
+      AuthorizationResult.ALLOWED,
+      AuthorizationResult.DENIED,
+      AuthorizationResult.DENIED,
+      AuthorizationResult.ALLOWED,
+      AuthorizationResult.ALLOWED,
+      AuthorizationResult.ALLOWED).asJava)
+    opaAuthorizer.getCache.size should be (4)
   }
 
   "OpaAuthorizer" should "not authorize when username does not match name of topic" in {
     val opaAuthorizer = setupAuthorizer()
-    val request = createRequest("alice-producer", "bob-topic", Write)
-    val input = request.input
+    val actions = List(
+      createAction("bob-topic", AclOperation.WRITE),
+    )
+    val request = createRequest("alice-producer", actions)
 
-    opaAuthorizer.authorize(input.session, input.operation, input.resource) should be (false)
+    opaAuthorizer.authorize(request.requestContext, request.actions.asJava) should be (List(AuthorizationResult.DENIED).asJava)
     opaAuthorizer.getCache.size should be (1)
   }
 
   "OpaAuthorizer" should "not authorize read request for producer" in {
     val opaAuthorizer = setupAuthorizer()
-    val request = createRequest("alice-producer", "alice-topic", Read)
-    val input = request.input
+    val actions = List(
+      createAction("alice-topic", AclOperation.READ),
+    )
+    val request = createRequest("alice-producer", actions)
 
-    opaAuthorizer.authorize(input.session, input.operation, input.resource) should be (false)
+    opaAuthorizer.authorize(request.requestContext, request.actions.asJava) should be (List(AuthorizationResult.DENIED).asJava)
     opaAuthorizer.getCache.size should be (1)
   }
 
   "OpaAuthorizer" should "not authorize write request for consumer" in {
     val opaAuthorizer = setupAuthorizer()
-    val request = createRequest("alice-consumer", "alice-topic", Write)
-    val input = request.input
+    val actions = List(
+      createAction("alice-topic", AclOperation.WRITE),
+    )
+    val request = createRequest("alice-consumer", actions)
 
-    opaAuthorizer.authorize(input.session, input.operation, input.resource) should be (false)
+    opaAuthorizer.authorize(request.requestContext, request.actions.asJava) should be (List(AuthorizationResult.DENIED).asJava)
     opaAuthorizer.getCache.size should be (1)
   }
 
   "OpaAuthorizer" should "cache the first request" in {
     val opaAuthorizer = setupAuthorizer()
-    val request = createRequest("alice-consumer", "alice-topic", Read)
-    val input = request.input
+    val actions = List(
+      createAction("alice-topic", AclOperation.READ),
+    )
+    val request = createRequest("alice-consumer", actions)
 
     for (_ <- 1 until 5) {
-      opaAuthorizer.authorize(input.session, input.operation, input.resource) should be (true)
+      opaAuthorizer.authorize(request.requestContext, request.actions.asJava) should be (List(AuthorizationResult.ALLOWED).asJava)
     }
 
     opaAuthorizer.getCache.size should be (1)
 
-    val otherRequest = createRequest("bob-consumer", "bob-topic", Read)
-    val nextInput = otherRequest.input
+    val otherActions = List(
+      createAction("bob-topic", AclOperation.READ),
+    )
+    val otherRequest = createRequest("bob-consumer", otherActions)
 
     for (_ <- 1 until 5) {
-      opaAuthorizer.authorize(nextInput.session, nextInput.operation, nextInput.resource) should be(true)
+      opaAuthorizer.authorize(otherRequest.requestContext, otherRequest.actions.asJava) should be (List(AuthorizationResult.ALLOWED).asJava)
     }
 
     opaAuthorizer.getCache.size should be (2)
@@ -101,27 +160,35 @@ class OpaAuthorizerSpec extends FlatSpec with Matchers with PrivateMethodTester 
 
   "OpaAuthorizer" should "not cache decisions while errors occur" in {
     val opaAuthorizer = setupAuthorizer("http://localhost/broken")
-    val request = createRequest("alice-consumer", "alice-topic", Write)
-    val input = request.input
+    val actions = List(
+      createAction("alice-topic", AclOperation.WRITE),
+    )
+    val request = createRequest("alice-consumer", actions)
 
-    opaAuthorizer.authorize(input.session, input.operation, input.resource) should be (false)
+    opaAuthorizer.authorize(request.requestContext, request.actions.asJava) should be (List(AuthorizationResult.DENIED).asJava)
     opaAuthorizer.getCache.size should be (0)
   }
 
   "OpaAuthorizer" should "authorize super users without checking with OPA" in {
     val opaAuthorizer = setupAuthorizer(opaUrl)
 
-    val resource = Resource(Topic, "alice-topic", PatternType.LITERAL)
-    val operation = Write
+    val actions1 = List(
+      createAction("alice-topic", AclOperation.WRITE),
+    )
+    val request1 = createRequest("CN=my-user", actions1)
+    opaAuthorizer.authorize(request1.requestContext, request1.actions.asJava) should be (List(AuthorizationResult.ALLOWED).asJava)
 
-    val session1 = Session(new KafkaPrincipal("User", "CN=my-user"), InetAddress.getLoopbackAddress)
-    opaAuthorizer.authorize(session1, operation, resource) should be (true)
+    val actions2 = List(
+      createAction("alice-topic", AclOperation.WRITE),
+    )
+    val request2 = createRequest("CN=my-user2,O=my-org", actions2)
+    opaAuthorizer.authorize(request2.requestContext, request2.actions.asJava) should be (List(AuthorizationResult.ALLOWED).asJava)
 
-    val session2 = Session(new KafkaPrincipal("User", "CN=my-user2,O=my-org"), InetAddress.getLoopbackAddress)
-    opaAuthorizer.authorize(session2, operation, resource) should be (true)
-
-    val session3 = Session(new KafkaPrincipal("User", "CN=my-user3"), InetAddress.getLoopbackAddress)
-    opaAuthorizer.authorize(session3, operation, resource) should be (false)
+    val actions3 = List(
+      createAction("alice-topic", AclOperation.WRITE),
+    )
+    val request3 = createRequest("CN=my-user3", actions3)
+    opaAuthorizer.authorize(request3.requestContext, request3.actions.asJava) should be (List(AuthorizationResult.DENIED).asJava)
   }
 
   def setupAuthorizer(url: String = opaUrl): OpaAuthorizer = {
@@ -134,11 +201,25 @@ class OpaAuthorizerSpec extends FlatSpec with Matchers with PrivateMethodTester 
     opaAuthorizer
   }
 
-  def createRequest(username: String, topic: String, operation: Operation): Request = {
-    val session = Session(new KafkaPrincipal("User", username), InetAddress.getLoopbackAddress)
-    val resource = Resource(Topic, topic, PatternType.LITERAL)
+  def createRequest(username: String, actions: List[Action]): FullRequest = {
+    val principal = new KafkaPrincipal("User", username)
+    val session = RequestChannel.Session(principal, InetAddress.getLoopbackAddress)
+    val authzReqContext = new AzRequestContext(
+      clientId = "rdkafka",
+      requestType = 1,
+      listenerName = "SASL_PLAINTEXT",
+      clientAddress = session.clientAddress,
+      principal = session.principal,
+      securityProtocol = SecurityProtocol.SASL_PLAINTEXT,
+      correlationId = new scala.util.Random().nextInt(1000),
+      requestVersion = 4)
 
-    Request(Input(session, operation, resource))
+    FullRequest(authzReqContext, actions)
+  }
+
+  def createAction(topic: String, operation: AclOperation): Action = {
+    val resource = new ResourcePattern(ResourceType.TOPIC, topic, PatternType.LITERAL)
+    new Action(operation, resource, 1, true, true)
   }
 
   def testOpaConnection(): Option[HttpResponse[String]] = {
@@ -150,3 +231,5 @@ class OpaAuthorizerSpec extends FlatSpec with Matchers with PrivateMethodTester 
     }
   }
 }
+
+case class FullRequest(requestContext: AzRequestContext, actions: List[Action])